crewai-system/requirements.txt
fastapi
uvicorn
pymongo
requests
beautifulsoup4
pandas
python-dotenv
praw
sentence-transformers
pinecone-client
langchain
fpdf
matplotlib
flask
flask-cors
transformers
torch
PyJWT
bcrypt